Moves transform_like protocol to protocols package
