--- src/pyrite/_types/transform.py
@@ -1,29 +1,0 @@
-from __future__ import annotations
-
-from typing import Protocol, runtime_checkable, TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from pygame import Vector2
-    from pygame.typing import Point
-
-
-@runtime_checkable
-class TransformLike(Protocol):
-
-    @property
-    def position(self) -> Vector2: ...
-
-    @position.setter
-    def position(self, position: Point): ...
-
-    @property
-    def rotation(self) -> float: ...
-
-    @rotation.setter
-    def rotation(self, rotation: float): ...
-
-    @property
-    def scale(self) -> Vector2: ...
-
-    @scale.setter
-    def scale(self, scale: Point): ...
